Move commands in system; Use same method for mail test command as the rest
