+++ api/commands/__init__.py
@@ -3,3 +3,4 @@ from .db_datas import db_datas
 from .users import users
 from .roles import roles
 from .tracks import tracks
+from .system import system
